Consertado o bug da funcao table_with_avgs que fazia a divisao entre um decimal e um float

+++ project/rsi.py
@@ -60,7 +60,7 @@ def table_with_first_rsi(data, period, rsi_g, rsi_l):
 
 
 def calc_avgs(last_rsi, actual_diff, period):
-    rs = actual_diff + (last_rsi * (period - 1)) / period
+    rs = actual_diff + Decimal((last_rsi * (period - 1)) / period)
     return rs
